面板数据结构变更
