+++ models/index.js
@@ -2,10 +2,8 @@ import Base from './Base.js'
 import Character from './Character.js'
 import Artifact from './Artifact.js'
 import ArtifactSet from './ArtifactSet.js'
-import Avatar from './Avatar.js'
 import AvatarData from './AvatarData.js'
 import AvatarArtis from './AvatarArtis.js'
-import AvatarList from './AvatarList.js'
 import Abyss from './Abyss.js'
 import Player from './Player.js'
 import ProfileServ from './ProfileServ.js'
@@ -26,9 +24,7 @@ export {
   Character,
   Artifact,
   ArtifactSet,
-  Avatar,
   AvatarData,
-  AvatarList,
   AvatarArtis,
   ProfileServ,
   ProfileReq,
