+++ resources/character/profile-detail.html
@@ -56,7 +56,7 @@
 
   <div class="data-info">
     {{if data.dataSource}}
-    <span>数据源：{{ {miao:"喵喵Api",enka:"Enka.Network",change:"面板变换"}[data.dataSource]||data.dataSource }}</span>
+    <span>数据源：{{ {change:"面板变换"}[data.dataSource]||data.dataSource }}</span>
     {{/if}}
     {{if data.updateTime}}
     <span class="time">{{data.updateTime}}</span>
